리팩토링: AITConvertCore 빌드 취소 상태를 AITBuildCancellation으로 분리
god-class AITConvertCore 분해(#36)의 첫 단계. 빌드 취소 상태
(취소 플래그/진행 중 비동기 작업 핸들/병렬 pnpm install 컨텍스트)와
관련 로직을 전용 정적 클래스 AITBuildCancellation으로 추출.

- 3개 volatile 정적 필드 + CancelBuild/ResetCancellation/IsCancelled/
  HasRunningAsyncTask/SetCurrentAsyncTask 본문을 새 클래스로 이동
- AITConvertCore는 기존 공개 API를 위임 래퍼로 유지 → 외부 호출자
  (AITPackageBuilder/AITPackageInitializer/AITNpmRunner/AITDeployManager)
  변경 없음
- 파이프라인 내 직접 필드 쓰기 5곳은 SetCurrentEarlyContext/
  SetCurrentAsyncTask 접근자로 라우팅
- volatile 의미·Debug.Log 메시지·정적 생성자 보존 (동작 불변)
