+++ Editor/AITBuildCancellation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5658cd61a6e48f09ec9f1a128811bbc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
